Refactor authentication context and integrate chat context; update related components and add utility files for animations and loading indicators.
