+++ NLP/fakultetski_sadržaj.txt
@@ -182,5 +182,9 @@
 
 - Cilj im je pružiti rješenja visoke tehnologije po mjeri krajnjih korisnika, a glavna djelatnost je razvoj softvera po narudžbi
 
-- Nasa lokacija je Zmaja od Bosne 7 
+- Nasa lokacija je Zmaja od Bosne 7
+
+
+
+Osnivac ove stranica i tvorac i jedan od studenata ipia akademija sam ja Amar Tubic
