+++ frontend/src/components/Header/Header.tsx
@@ -2,70 +2,47 @@ import { useState, useEffect } from 'react'
 import { motion, AnimatePresence } from 'framer-motion'
 import { GiHamburgerMenu } from 'react-icons/gi'
 import { IoClose } from 'react-icons/io5'
-import { useChat } from '../../Context'
+import { useAuth } from '../../Context'
 import { Link, useNavigate, useLocation } from 'react-router-dom'
 
+import {
+    sidebarVariants,
+    dropdownVariants,
+    headerVariants,
+    navLinks,
+} from './AnimationHeader'
+
+// Header components
 const Header = () => {
     const [isSidebarOpen, setIsSidebarOpen] = useState(false)
     const [isOpen, setIsOpen] = useState(false)
     const [scrolled, setScrolled] = useState(false)
     const nav = useNavigate()
     const location = useLocation()
-    const { studentMail, logout } = useChat()
+    const { studentMail, logout } = useAuth()
 
-    // Handle scroll effect
     useEffect(() => {
-        const handleScroll = () => {
-            setScrolled(window.scrollY > 20)
+        const debounce = (func, delay) => {
+            let timeout
+            return (...args) => {
+                clearTimeout(timeout)
+                timeout = setTimeout(() => func(...args), delay)
+            }
         }
+
+        const handleScroll = debounce(() => {
+            setScrolled(window.scrollY > 20)
+        }, 100) // 100ms debounce
+
         window.addEventListener('scroll', handleScroll)
-        return () => window.removeEventListener('scroll', handleScroll)
+
+        return () => {
+            window.removeEventListener('scroll', handleScroll)
+        }
     }, [])
 
     const toggleMenu = () => setIsOpen(!isOpen)
     const toggleSidebar = () => setIsSidebarOpen(!isSidebarOpen)
-
-    const navLinks = [
-        { to: '/', text: 'Početna' },
-        { to: '/about', text: 'O nama' },
-        { to: '/programs', text: 'Studijski programi' },
-        { to: '/news', text: 'Novosti' },
-        { to: '/contact', text: 'Kontakt' },
-    ]
-
-    const headerVariants = {
-        initial: { y: -100 },
-        animate: {
-            y: 0,
-            transition: { type: 'spring', stiffness: 100, damping: 20 },
-        },
-    }
-
-    const dropdownVariants = {
-        hidden: { opacity: 0, y: -20 },
-        visible: {
-            opacity: 1,
-            y: 0,
-            transition: { type: 'spring', stiffness: 400, damping: 30 },
-        },
-        exit: {
-            opacity: 0,
-            y: -20,
-            transition: { duration: 0.2 },
-        },
-    }
-
-    const sidebarVariants = {
-        hidden: { x: '100%' },
-        visible: {
-            x: 0,
-            transition: { type: 'spring', stiffness: 300, damping: 30 },
-        },
-        exit: {
-            x: '100%',
-            transition: { duration: 0.3 },
-        },
-    }
 
     return (
         <motion.header
