+++ frontend/src/types/user.ts
@@ -1,0 +1,9 @@
+export interface UserDetails {
+    ime: string
+    prezime: string
+    email: string
+    tipUsera: string
+    indeks?: string
+    godinaStudija?: string
+    smjerStudija?: string
+}
